@tczajka 2020-10-23
Some benchmarks.

Before:
```
test divide_0             ... bench:         707 ns/iter (+/- 32)
test divide_1             ... bench:      18,700 ns/iter (+/- 20)
test divide_2             ... bench:     431,555 ns/iter (+/- 33,198)
test divide_big_little    ... bench:      23,667 ns/iter (+/- 23)
```
After:
```
test divide_0             ... bench:         706 ns/iter (+/- 1)
test divide_1             ... bench:      14,276 ns/iter (+/- 20)
test divide_2             ... bench:     307,589 ns/iter (+/- 569)
test divide_big_little    ... bench:      23,665 ns/iter (+/- 25)
```

Before:
```
test remainder_0          ... bench:         715 ns/iter (+/- 14)
test remainder_1          ... bench:      18,733 ns/iter (+/- 48)
test remainder_2          ... bench:     436,301 ns/iter (+/- 9,878)
test remainder_big_little ... bench:      23,477 ns/iter (+/- 32)
```
After:
```
test remainder_0          ... bench:         712 ns/iter (+/- 13)
test remainder_1          ... bench:      14,260 ns/iter (+/- 15)
test remainder_2          ... bench:     307,634 ns/iter (+/- 585)
test remainder_big_little ... bench:      23,477 ns/iter (+/- 16)
```

Before:
```
test gcd_euclid_0064 ... bench:       2,719 ns/iter (+/- 50)
test gcd_euclid_0256 ... bench:      43,623 ns/iter (+/- 149)
test gcd_euclid_1024 ... bench:     248,941 ns/iter (+/- 613)
test gcd_euclid_4096 ... bench:   1,694,518 ns/iter (+/- 2,088)
```
After:
```
test gcd_euclid_0064 ... bench:       2,757 ns/iter (+/- 34)
test gcd_euclid_0256 ... bench:      28,305 ns/iter (+/- 147)
test gcd_euclid_1024 ... bench:     159,789 ns/iter (+/- 527)
test gcd_euclid_4096 ... bench:   1,180,507 ns/iter (+/- 6,867)
```

Before:
```
test big2k_sqrt      ... bench:      32,807 ns/iter (+/- 316)
test big4k_cbrt      ... bench:      52,402 ns/iter (+/- 128)
test big4k_nth_10    ... bench:      29,072 ns/iter (+/- 343)
test big4k_nth_100   ... bench:      15,317 ns/iter (+/- 599)
test big4k_nth_1000  ... bench:      58,758 ns/iter (+/- 602)
test big4k_nth_10000 ... bench:          31 ns/iter (+/- 0)
test big4k_sqrt      ... bench:      66,525 ns/iter (+/- 588)
```
After:
```
test big2k_sqrt      ... bench:      23,926 ns/iter (+/- 169)
test big4k_cbrt      ... bench:      38,278 ns/iter (+/- 480)
test big4k_nth_10    ... bench:      22,307 ns/iter (+/- 162)
test big4k_nth_100   ... bench:      14,299 ns/iter (+/- 41)
test big4k_nth_1000  ... bench:      57,104 ns/iter (+/- 117)
test big4k_nth_10000 ... bench:          31 ns/iter (+/- 1)
test big4k_sqrt      ... bench:      47,867 ns/iter (+/- 61)
```